+++ CHANGELOG.md
@@ -5,7 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## dev
 
+* Enhancements
+  * [Core] Support images of other repositories beyond Docker Hub;
+
 * Bug
+  * [Core] Simplification the images downloads process;
   * [Core] Fixing regex bug in expand envs (not expand numbers: `$1` or `${2}`);
   * [Core] Adding image envs in expand envs variable process;
   * [Core] Adding support to escape variables in command options (shell and system);